css/css-transitions/before-DOMContentLoaded-001.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transitions Test: Transitioning before DOMContentLoaded event</title>
<meta name="assert" content="Test checks that transitions are run even before the document has left parsing mode">
<link rel="help" title="5. Transition Events" href="http://www.w3.org/TR/css3-transitions/#transition-events">
<link rel="author" title="Rodney Rehm" href="http://rodneyrehm.de/en/">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./support/helper.js"></script>

</head>
<body>

<script>
// Make sure a transition is run before DOMContentLoaded
var isDOMContentLoaded = false;
window.addEventListener('DOMContentLoaded', function() {
  isDOMContentLoaded = true;
}, false);

async_test(t => {
  const div = addDiv(t, { style: 'transition: height .01s linear; ' +
                                 'height: 1px' });
  getComputedStyle(div).height;
  div.style.height = '100px';

  div.addEventListener('transitionend', t.step_func_done(() => {
    assert_false(isDOMContentLoaded,
      'DOMContentLoaded may not have happened yet');
  }));
});
</script>

<!-- Delay DOMContentLoaded long enough for transition to run -->
<script src="support/empty.js?pipe=trickle(d3)" id="delay"></script>
</body>
</html>
